Move sidepanel funcs to separate file
